move binning tuple utils to pmacc
