+++ include/picongpu/plugins/binning/utility.hpp
@@ -19,8 +19,7 @@
 
 #pragma once
 
-#include <pmacc/attribute/FunctionSpecifier.hpp>
-#include <pmacc/memory/STLTuple.hpp>
+#include <pmacc/memory/tuple/utility.hpp>
 
 #include <tuple>
 #include <utility>
@@ -29,71 +28,6 @@ namespace picongpu
 {
     namespace plugins::binning
     {
-        namespace detail
-        {
-            template<typename TFunc, typename TPmaccTuple, std::size_t... Is>
-            HDINLINE constexpr decltype(auto) applyImpl(TFunc&& f, TPmaccTuple&& t, std::index_sequence<Is...>)
-            {
-                return std::forward<TFunc>(f)(pmacc::memory::tuple::get<Is>(std::forward<TPmaccTuple>(t))...);
-            }
-
-            template<typename TFunc, typename TPmaccTuple, std::size_t... Is>
-            HDINLINE constexpr decltype(auto) applyEnumerateImpl(
-                TFunc&& f,
-                TPmaccTuple&& t,
-                std::index_sequence<Is...>)
-            {
-                return std::forward<TFunc>(f)(pmacc::memory::tuple::make_tuple(
-                    std::integral_constant<std::size_t, Is>{},
-                    pmacc::memory::tuple::get<Is>(std::forward<TPmaccTuple>(t)))...);
-            }
-
-        } // namespace detail
-
-        // takes pmacc::memory::tuple::Tuple
-        template<typename TFunc, typename TPmaccTuple>
-        HDINLINE constexpr decltype(auto) apply(TFunc&& f, TPmaccTuple&& t)
-        {
-            return detail::applyImpl(
-                std::forward<TFunc>(f),
-                std::forward<TPmaccTuple>(t),
-                std::make_index_sequence<pmacc::memory::tuple::tuple_size_v<TPmaccTuple>>{});
-        }
-
-        // takes pmacc::memory::tuple::Tuple
-        template<typename TFunc, typename TPmaccTuple>
-        HDINLINE constexpr decltype(auto) applyEnumerate(TFunc&& f, TPmaccTuple&& t)
-        {
-            return detail::applyEnumerateImpl(
-                std::forward<TFunc>(f),
-                std::forward<TPmaccTuple>(t),
-                std::make_index_sequence<pmacc::memory::tuple::tuple_size_v<TPmaccTuple>>{});
-        }
-
-        namespace detail
-        {
-            template<size_t... Is, typename... Args, typename Functor>
-            constexpr auto tupleMapHelper(
-                std::index_sequence<Is...>,
-                std::tuple<Args...> const& tuple,
-                Functor&& functor)
-            {
-                return pmacc::memory::tuple::make_tuple(std::forward<Functor>(functor)(std::get<Is>(tuple))...);
-            }
-        } // namespace detail
-
-        /**
-         * @brief create a new tuple from the return value of a functor applied on all arguments of a tuple
-         */
-        template<typename... Args, typename Functor>
-        constexpr auto tupleMap(std::tuple<Args...> const& tuple, Functor&& functor)
-        {
-            return detail::tupleMapHelper(
-                std::make_index_sequence<sizeof...(Args)>{},
-                tuple,
-                std::forward<Functor>(functor));
-        }
-
         template<typename... Args>
         constexpr auto createTuple(Args&&... args)
         {
